FIX : BDBD-412 게시글 관련 문제점 수정

+++ src/main/java/com/fakedevelopers/bidderbidder/dto/ProductWriteDto.java
@@ -23,7 +23,7 @@ public class ProductWriteDto {
   private final long openingBid;
 
   @Min(1)
-  private final long tick;
+  private final int tick;
 
   private final Long hopePrice;
